requirements.txt
pytz>=2019.3
pyxdg
tasklib>=1.3.0
tzlocal>=2.0.0
urwid>=2.1.0
